Replace require 'fs' with fs.coffee exporting graceful-fs
Since this module always deal with lots and lots of static files
I believe it is a good idea to add a central point where all fs
calls can be traced/inspect/modifed/patched/whatever.

In this commit I'm simply replacing all `require 'fs'` found in
the codebase with require `require './fs.coffee' which now
simply exports the `graceful-fs` module.

This is to combat the EMFILE issue:
https://github.com/techpines/asset-rack/issues/94