Express 4
Now that Express 4 no longer excepts objects as middleware we have to
use callbacks instead.

To keep asset-rack working in express 4, with some kind of fallback
capabilities for express 3, I'm proposing to specify the `Rack#handle`
method when adding the rack/asset as middleware. I've also upgraded
CoffeeScript, so I could easily bind the class's method to any instance
of it with the new fat-arrow syntax.

Another problem is the `express.Response#locals` property is no longer a
function. Therefore I've added the assets' "locals" property directly to
the express instance.